fix(mig123): cast avg_accuracy to numeric for ROUND (accuracy is REAL)
quiz_sessions.accuracy is REAL, so AVG() returns double precision, and Postgres
has no ROUND(double precision, int) — only ROUND(numeric, int). Applying mig 123
failed with 'function round(double precision, integer) does not exist'. Cast the
average to numeric before rounding.

+++ backend/migrations/123_quiz_admin_student_rollup.sql
@@ -50,7 +50,9 @@ LANGUAGE sql STABLE AS $$
            sess.sessions,
            sess.graded_sessions,
            sess.total_time_sec,
-           ROUND(sess.avg_accuracy, 3)              AS avg_accuracy,
+           -- accuracy is REAL, so AVG() is double precision; ROUND(double, int)
+           -- doesn't exist in Postgres — cast to numeric for the 2-arg ROUND.
+           ROUND(sess.avg_accuracy::numeric, 3)     AS avg_accuracy,
            COALESCE(mastered.words_mastered, 0)     AS words_mastered,
            sess.last_active
     FROM sess
